style(ui): adjust design of IO and message view components
