+++ web/src/pages/project/[projectId]/datasets/[datasetId]/compare.tsx
@@ -13,7 +13,7 @@ import {
   PopoverTrigger,
   PopoverContent,
 } from "@/src/components/ui/popover";
-import { MarkdownOrJsonView } from "@/src/components/trace/IOPreview";
+import { MarkdownJsonView } from "@/src/components/ui/MarkdownJsonView";
 import {
   Dialog,
   DialogContent,
@@ -223,9 +223,7 @@ export default function DatasetCompare() {
                 </div>
                 <div>
                   <h4 className="mb-1 font-medium">Metadata</h4>
-                  <MarkdownOrJsonView
-                    content={dataset.data?.metadata ?? null}
-                  />
+                  <MarkdownJsonView content={dataset.data?.metadata ?? null} />
                 </div>
               </div>
             </PopoverContent>
